+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import "./App.css";
-import Main from "./main";
+import Main from "./main/main.js";
 
 class App extends Component {
   render() {
