+++ client/src/main/pages/home/home.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import "./index.css";
+import "./home.css";
 
 class Home extends Component {
   state = {};
